batch: Move line sequence search into package
The line-matching package owns exact sequence equality while contextual
sequence search remains at the batch package root.

Placement code uses search as an extension of exact comparison, so the
current separation hides the helper dependency behind an unrelated
directory boundary.

This commit moves exact sequence search beside equality and keeps its
selected import path available until focused tests migrate.

Subsequent commits will move range views and source lineage before the
package transition removes its old paths.
